extract abstrct `PerimeterxRiskClient`
